[BE] :recycle: 폴더명 리팩토링
