+++ server/DevSquad/src/main/java/WOOMOOL/DevSquad/member/dto/MemberProfileDto.java
@@ -1,9 +1,8 @@
 package WOOMOOL.DevSquad.member.dto;
 
 
-import WOOMOOL.DevSquad.position.entity.Position;
-import WOOMOOL.DevSquad.projectBoard.entity.Project;
-import WOOMOOL.DevSquad.studyBoard.entity.Study;
+import WOOMOOL.DevSquad.projectboard.entity.Project;
+import WOOMOOL.DevSquad.studyboard.entity.Study;
 import lombok.*;
 
 import javax.validation.constraints.NotBlank;
