+++ server/DevSquad/src/main/java/WOOMOOL/DevSquad/member/entity/MemberProfile.java
@@ -1,14 +1,12 @@
 package WOOMOOL.DevSquad.member.entity;
 
 import WOOMOOL.DevSquad.position.entity.Position;
-import WOOMOOL.DevSquad.projectBoard.entity.Project;
-import WOOMOOL.DevSquad.studyBoard.entity.Study;
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import WOOMOOL.DevSquad.projectboard.entity.Project;
+import WOOMOOL.DevSquad.studyboard.entity.Study;
 import lombok.*;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
