+++ server/DevSquad/src/main/java/WOOMOOL/DevSquad/projectboard/entity/Project.java
@@ -1,7 +1,7 @@
-package WOOMOOL.DevSquad.studyBoard.entity;
+package WOOMOOL.DevSquad.projectboard.entity;
 
-import WOOMOOL.DevSquad.member.entity.MemberProfile;
 import WOOMOOL.DevSquad.board.entity.Board;
+import WOOMOOL.DevSquad.member.entity.MemberProfile;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,13 +10,21 @@ import java.time.LocalDateTime;
 
 @Entity
 @Getter @Setter
-public class Study extends Board {
+public class Project extends Board {
 
     @Column(nullable = false)
     private String title;
 
     @Column(nullable = false)
     private String content;
+
+    // 미정 / 날짜 형식
+    @Column(nullable = false)
+    private String startDate;
+
+    // 미정 / 날짜 형식
+    @Column(nullable = false)
+    private String deadline;
 
     @Column(nullable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
@@ -36,20 +44,20 @@ public class Study extends Board {
     // Soft Delete
     @Enumerated(value = EnumType.STRING)
     @Column(nullable = false)
-    private Study.StudyStatus studyStatus = Study.StudyStatus.STUDY_POSTED;
+    private ProjectStatus projectStatus = ProjectStatus.PROJECT_POSTED;
 
     @ManyToOne  // N : 1
     @JoinColumn(name = "MEMBER_PROFILE_ID")
     private MemberProfile memberProfile;
 
-    public enum StudyStatus {
-        STUDY_POSTED("게시 중"),
-        STUDY_CLOSED("모집 완료"),
-        STUDY_DELETED("삭제");
+    public enum ProjectStatus {
+        PROJECT_POSTED("게시 중"),
+        PROJECT_CLOSED("모집 완료"),
+        PROJECT_DELETED("삭제");
 
         private String status;
 
-        StudyStatus(String status) {
+        ProjectStatus(String status) {
             this.status = status;
         }
     }
